Change src/core/database to a modular plugin
