+++ src/shared/user/model.js
@@ -1,8 +1,10 @@
 'use strict';
 
+const Promise = require('bluebird');
+
 const bcrypt = require('bcryptjs');
 const shortid = require('shortid');
-const mongoose = require('mongoose');
+const mongoose = require('k7-mongoose').mongoose;
 
 const Schema = new mongoose.Schema({
   name: {
@@ -55,7 +57,7 @@ Schema.methods.validatePassword = function (requestPassword) {
 
 const UserModel = mongoose.model('User', Schema);
 
-module.exports = UserModel;
+module.exports = Promise.promisifyAll(UserModel);
 
 function hashPassword (password) {
   if (!password) {
